requirements.txt
# Core simulation
gymnasium>=0.26.0
pybullet>=3.2.5
numpy>=1.21.0
pyyaml>=6.0

# Optimal assignment (Hungarian algorithm)
scipy>=1.7.0

# JIT acceleration (optional -- graceful fallback if absent)
numba>=0.53

# Local panda-gym fork (install with: pip install -e ./panda-gym)
# panda-gym

# Testing
pytest>=7.0.0
